+++ src/app/components/tasks-document-export/task-form/task-document-export-form.component.ts
@@ -262,8 +262,8 @@ export class TaskDocumentExportFormComponent extends BaseFormComponent<TaskProje
       taskAvailabilityService: this.taskAvailabilityService,
       taskService: this.taskService,
       isNew: () => this.isNew(),
-      entity: this.entityToEdit,
-      entityId: this.entityID,
+      entity: () => this.entityToEdit,
+      entityId: () => this.entityID,
     };
   }
 }