+++ src/app/components/tasks-map-image/task-form/task-map-image-form.component.ts
@@ -701,8 +701,8 @@ export class TaskMapImageFormComponent extends BaseFormComponent<TaskProjection>
       taskAvailabilityService: this.taskAvailabilityService,
       taskService: this.taskService,
       isNew: () => this.isNew(),
-      entity: this.entityToEdit,
-      entityId: this.entityID,
+      entity: () => this.entityToEdit,
+      entityId: () => this.entityID,
     };
   }